+++ IL2-SR-CommonTests/Network/IL2RadioSyncHandlerTests.cs
@@ -1,6 +1,8 @@
 using Ciribob.IL2.SimpleRadio.Standalone.Client.Network.IL2;
 using Ciribob.IL2.SimpleRadio.Standalone.Common;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System.Net;
+using System.Net.Sockets;
 
 namespace Ciribob.IL2.SimpleRadio.Standalone.Common.Tests.Network
 {
@@ -47,6 +49,33 @@ namespace Ciribob.IL2.SimpleRadio.Standalone.Common.Tests.Network
             var changed = IL2RadioSyncHandler.ApplyControlData(state, -1, 1);
 
             Assert.IsFalse(changed);
+        }
+
+        [TestMethod]
+        public void TelemetryListenerOwnsItsPortExclusively()
+        {
+            using (UdpClient listener = IL2RadioSyncHandler.BindExclusiveListener(
+                       new IPEndPoint(IPAddress.Loopback, 0)))
+            {
+                Assert.IsTrue(listener.ExclusiveAddressUse);
+                int port = ((IPEndPoint)listener.Client.LocalEndPoint).Port;
+
+                using (UdpClient competingListener = new UdpClient())
+                {
+                    competingListener.Client.SetSocketOption(
+                        SocketOptionLevel.Socket,
+                        SocketOptionName.ReuseAddress,
+                        true);
+                    try
+                    {
+                        competingListener.Client.Bind(new IPEndPoint(IPAddress.Loopback, port));
+                        Assert.Fail("A second listener must not be able to take over the SRS telemetry port.");
+                    }
+                    catch (SocketException)
+                    {
+                    }
+                }
+            }
         }
     }
 }